Added files
Removed init functions from main, and put in separate files.
