+++ inc/ADC.h
@@ -1,0 +1,1 @@
+extern void ADC_init();
