+++ inc/DMA.h
@@ -1,0 +1,1 @@
+extern void DMA_init();
