+++ inc/UART.h
@@ -1,0 +1,1 @@
+extern void UART_init();
