+++ src/UART.c
@@ -1,0 +1,68 @@
+#include "stm32f4xx_usart.h"
+
+static void Configure_USART2(void){
+	RCC_APB2PeriphClockCmd(RCC_APB1Periph_USART2, ENABLE);
+	/**
+	 * Set Baudrate to value you pass to function
+	 * Disable Hardware Flow control
+	 * Set Mode To TX and RX, so USART will work in full-duplex mode
+	 * Disable parity bit
+	 * Set 1 stop bit
+	 * Set Data bits to 8
+	 *
+	 * Initialize USART2
+	 * Activate USART2
+	 */
+	USART_InitTypeDef USART_InitStruct;
+
+	USART_InitStruct.USART_BaudRate = 9600;
+	USART_InitStruct.USART_HardwareFlowControl = USART_HardwareFlowControl_None;
+	USART_InitStruct.USART_Mode = USART_Mode_Tx | USART_Mode_Rx;
+	USART_InitStruct.USART_Parity = USART_Parity_No;
+	USART_InitStruct.USART_StopBits = USART_StopBits_1;
+	USART_InitStruct.USART_WordLength = USART_WordLength_8b;
+	USART_Init(USART2, &USART_InitStruct);
+
+	USART_ClockInitTypeDef USART_ClockInitStruct;
+	USART_ClockInitStruct.USART_Clock = USART_Clock_Disable;
+	USART_ClockInitStruct.USART_CPOL = USART_CPOL_Low;
+	USART_ClockInitStruct.USART_CPHA = USART_CPHA_2Edge;
+	USART_ClockInitStruct.USART_LastBit = USART_LastBit_Disable;
+	USART_ClockInit(USART2, &USART_ClockInitStruct);
+
+	USART_DMACmd(USART2, (USART_DMAReq_Rx | USART_DMAReq_Tx), ENABLE);
+	USART_Cmd(USART2, ENABLE);
+}
+
+static void Configure_GPIO_USART2(){
+	GPIO_InitTypeDef GPIO_InitStructure; //Variable used to setup the GPIO pins
+
+	//Using PD5 and PD6 instead for USART2
+	RCC_AHB1PeriphClockCmd(RCC_AHB1Periph_GPIOD, ENABLE);
+
+	GPIO_PinAFConfig(GPIOD, GPIO_PinSource5, GPIO_AF_USART2);
+	GPIO_PinAFConfig(GPIOD, GPIO_PinSource6, GPIO_AF_USART2);
+	// Initialize pins as alternating function
+	GPIO_InitStructure.GPIO_Pin = GPIO_Pin_5 | GPIO_Pin_6;
+	GPIO_InitStructure.GPIO_Mode = GPIO_Mode_AF;
+	GPIO_InitStructure.GPIO_OType = GPIO_OType_PP;
+	GPIO_InitStructure.GPIO_PuPd = GPIO_PuPd_UP;
+	GPIO_InitStructure.GPIO_Speed = GPIO_Speed_100MHz;
+	GPIO_Init(GPIOD, &GPIO_InitStructure);
+}
+
+static void Configure_NVIC_USART2(){
+	//Enable UART channel IRQ Channel */
+	NVIC_InitTypeDef NVIC_InitStructure;
+	NVIC_InitStructure.NVIC_IRQChannel = DMA2_Stream6_IRQn;// USART2_IRQn;
+	NVIC_InitStructure.NVIC_IRQChannelPreemptionPriority = 0;
+	NVIC_InitStructure.NVIC_IRQChannelSubPriority = 1;
+	NVIC_InitStructure.NVIC_IRQChannelCmd = ENABLE;
+	NVIC_Init(&NVIC_InitStructure);
+}
+
+extern void UART_init(){
+	Configure_GPIO_USART2();
+	Configure_NVIC_USART2();
+	Configure_USART2();
+}
