+++ src/main.c
@@ -9,281 +9,25 @@
 */
 
 #include "stm32f4xx.h"
-#include "stm32f4xx_adc.h"
-#include "stm32f4xx_usart.h"
+#include "UART.h"
+#include "DMA.h"
+#include "ADC.h"
 #include "main.h"
-#define ARRAYSIZE 8
-#define BaudRate 9600
-volatile uint16_t ADC_values[ARRAYSIZE];
-volatile uint16_t ADC_values2[ARRAYSIZE];
-int dma2_status = 0;
-
-void ADCInit(void);
-void DMA2Init(void); //peripheral to memory
-void NVICInit(void);
-void UARTInit(void);
-void DMA1Init(void); //memory to peripheral
-void GPIOInit(void);
 
 int main(void)
 {
-GPIOInit();
-ADCInit();
-DMA2Init();
-NVICInit();
-UARTInit();
-DMA1Init();
-//Start ADC1 Software Conversion
-ADC_SoftwareStartConv(ADC1);
+	ADC_init();
+	UART_init();
+	DMA_init();
+	//Start ADC1 Software Conversion
+	ADC_SoftwareStartConv(ADC1);
 
 
-while (1)
-  {
-
-  }
-}
-void GPIOInit(void){
-	/**
-	* Configuring GPIOA for ADC
-	*/
-	RCC_AHB1PeriphClockCmd(RCC_AHB1Periph_GPIOA, ENABLE);
-
-	GPIO_InitTypeDef GPIO_InitStructure; //Variable used to setup the GPIO pins
-	//==Configure ADC pins (PA0 -> Channel 0 to PA7 -> Channel 7) as analog inputs==
-	GPIO_StructInit(&GPIO_InitStructure); // Reset init structure, if not it can cause issues...
-	GPIO_InitStructure.GPIO_Pin = GPIO_Pin_0 | GPIO_Pin_1| GPIO_Pin_2| GPIO_Pin_3| GPIO_Pin_4| GPIO_Pin_5| GPIO_Pin_6| GPIO_Pin_7;
-	GPIO_InitStructure.GPIO_Mode = GPIO_Mode_AIN;
-	GPIO_InitStructure.GPIO_PuPd = GPIO_PuPd_DOWN;
-	GPIO_Init(GPIOA, &GPIO_InitStructure);
-
-	/**
-	* Configuring GPIOA for USART
-	*/
-//	GPIO_PinAFConfig(GPIOA, GPIO_PinSource9, GPIO_AF_USART2);
-//	GPIO_PinAFConfig(GPIOA, GPIO_PinSource10, GPIO_AF_USART2);
-//	// Initialize pins as alternating function
-//	GPIO_InitStructure.GPIO_Pin = GPIO_Pin_9 | GPIO_Pin_10;
-//	GPIO_InitStructure.GPIO_Mode = GPIO_Mode_AF;
-//	GPIO_InitStructure.GPIO_OType = GPIO_OType_PP;
-//	GPIO_InitStructure.GPIO_PuPd = GPIO_PuPd_UP;
-//	GPIO_InitStructure.GPIO_Speed = GPIO_Speed_100MHz;
-//	GPIO_Init(GPIOA, &GPIO_InitStructure);
-
-	//Using PD5 and PD6 instead for USART2
-	RCC_AHB1PeriphClockCmd(RCC_AHB1Periph_GPIOD, ENABLE);
-
-	GPIO_PinAFConfig(GPIOD, GPIO_PinSource5, GPIO_AF_USART2);
-	GPIO_PinAFConfig(GPIOD, GPIO_PinSource6, GPIO_AF_USART2);
-	// Initialize pins as alternating function
-	GPIO_InitStructure.GPIO_Pin = GPIO_Pin_5 | GPIO_Pin_6;
-	GPIO_InitStructure.GPIO_Mode = GPIO_Mode_AF;
-	GPIO_InitStructure.GPIO_OType = GPIO_OType_PP;
-	GPIO_InitStructure.GPIO_PuPd = GPIO_PuPd_UP;
-	GPIO_InitStructure.GPIO_Speed = GPIO_Speed_100MHz;
-	GPIO_Init(GPIOD, &GPIO_InitStructure);
-
-}
-
-void ADCInit(void){
-	/*enable peripheral clock for ADC1 */
-	RCC_APB2PeriphClockCmd(RCC_APB2Periph_ADC1, ENABLE);
-
-	ADC_InitTypeDef ADC_InitStructure;
-	//ADC1 configuration
-
-	//select continuous conversion mode
-	//ADC_InitStructure.ADC_Mode = ADC_Mode_Independent;
-
-	//We will convert multiple channels
-	ADC_InitStructure.ADC_ScanConvMode = ENABLE;
-	//we will convert one time
-	ADC_InitStructure.ADC_ContinuousConvMode = ENABLE;
-	//select no external triggering
-	ADC_InitStructure.ADC_ExternalTrigConvEdge = 0;
-	ADC_InitStructure.ADC_ExternalTrigConv = 0;
-	//right 12-bit data alignment in ADC data register
-	ADC_InitStructure.ADC_DataAlign = ADC_DataAlign_Right;
-	//8 ADC conversions for sequencer
-	ADC_InitStructure.ADC_NbrOfConversion = 8;
-
-	//load structure values to control and status registers
-	ADC_Init(ADC1, &ADC_InitStructure);
-
-	//channel configuration
-	ADC_RegularChannelConfig(ADC1, ADC_Channel_0, 1, ADC_SampleTime_56Cycles);
-	ADC_RegularChannelConfig(ADC1, ADC_Channel_1, 2, ADC_SampleTime_56Cycles);
-	ADC_RegularChannelConfig(ADC1, ADC_Channel_2, 3, ADC_SampleTime_56Cycles);
-	ADC_RegularChannelConfig(ADC1, ADC_Channel_3, 4, ADC_SampleTime_56Cycles);
-	ADC_RegularChannelConfig(ADC1, ADC_Channel_4, 5, ADC_SampleTime_56Cycles);
-	ADC_RegularChannelConfig(ADC1, ADC_Channel_5, 6, ADC_SampleTime_56Cycles);
-	ADC_RegularChannelConfig(ADC1, ADC_Channel_6, 7, ADC_SampleTime_56Cycles);
-	ADC_RegularChannelConfig(ADC1, ADC_Channel_7, 8, ADC_SampleTime_56Cycles);
-
-	ADC_CommonInitTypeDef ADC_CommonInitStructure;
-	ADC_CommonInitStructure.ADC_Mode = ADC_Mode_Independent;
-	ADC_CommonInitStructure.ADC_Prescaler = ADC_Prescaler_Div2; //appropriate speed?
-	ADC_CommonInitStructure.ADC_DMAAccessMode = ADC_DMAAccessMode_1;
-	ADC_CommonInitStructure.ADC_TwoSamplingDelay = ADC_TwoSamplingDelay_8Cycles;
-	ADC_CommonInit(&ADC_CommonInitStructure);
-
-	//Enables or disables the ADC DMA request after last transfer (Single-ADC mode)
-	ADC_DMARequestAfterLastTransferCmd(ADC1, ENABLE);
-	//Enable ADC1
-
-	ADC_Cmd(ADC1, ENABLE);
-	//enable DMA for ADC
-	ADC_DMACmd(ADC1, ENABLE);
-}
-void DMA2Init(void){
-	//enable DMA2 clock
-	RCC_AHB1PeriphClockCmd(RCC_AHB1Periph_DMA2, ENABLE);
-	//create DMA structure
-	DMA_InitTypeDef  DMA_InitStructure;
-	//reset DMA2 channe1 to default values;
-	DMA_DeInit(DMA2_Stream0);
-
-	DMA_InitStructure.DMA_Channel = DMA_Channel_0;
-	DMA_InitStructure.DMA_FIFOMode = DMA_FIFOMode_Disable;  //not using FIFO
-	DMA_InitStructure.DMA_FIFOThreshold = DMA_FIFOThreshold_HalfFull; //not using FIFO
-	DMA_InitStructure.DMA_MemoryBurst = DMA_MemoryBurst_Single;
-	DMA_InitStructure.DMA_PeripheralBurst = DMA_PeripheralBurst_Single;
-
-	//setting circular mode
-	DMA_InitStructure.DMA_Mode = DMA_Mode_Circular;
-	//medium priority
-	DMA_InitStructure.DMA_Priority = DMA_Priority_High;
-	//source and destination data size word=32bit
-	DMA_InitStructure.DMA_PeripheralDataSize = DMA_PeripheralDataSize_HalfWord;
-	DMA_InitStructure.DMA_MemoryDataSize = DMA_MemoryDataSize_HalfWord;
-	//automatic memory destination increment enable.
-	DMA_InitStructure.DMA_MemoryInc = DMA_MemoryInc_Enable;
-	//source address increment disable
-	DMA_InitStructure.DMA_PeripheralInc = DMA_PeripheralInc_Disable;
-	//Location assigned to peripheral register will be source
-	DMA_InitStructure.DMA_DIR = DMA_DIR_PeripheralToMemory;
-	//chunk of data to be transfered
-	DMA_InitStructure.DMA_BufferSize = ARRAYSIZE;
-	//source and destination start addresses
-	DMA_InitStructure.DMA_PeripheralBaseAddr = (uint32_t)&ADC1->DR;
-	DMA_InitStructure.DMA_Memory0BaseAddr = (uint32_t)&ADC_values[0];
-
-	//CONFIGURE DOUBLE BUFFER MODE, MEMORY0BASEADDR configured as current address
-	DMA_DoubleBufferModeConfig(DMA2_Stream0,&ADC_values2, DMA_Memory_0);
-	DMA_DoubleBufferModeCmd(DMA2_Stream0, ENABLE);
-
-	DMA_Init(DMA2_Stream0, &DMA_InitStructure);
-	// Enable DMA2 Channel Transfer Complete interrupt
-	DMA_ITConfig(DMA2_Stream0, DMA_IT_TC, ENABLE);
-	DMA_Cmd(DMA2_Stream0, ENABLE); //Enable the DMA2 - Stream 0
-}
-
-void DMA1Init(void) {
-	//enable DMA2 clock
-	RCC_AHB1PeriphClockCmd(RCC_AHB1Periph_DMA1, ENABLE);
-	//create DMA structure
-	DMA_InitTypeDef  DMA_InitStructure;
-	//reset DMA2 channel to default values;
-	DMA_DeInit(DMA1_Stream6); //stream 6 for usart2 tx
-
-	DMA_InitStructure.DMA_Channel = DMA_Channel_4;
-	DMA_InitStructure.DMA_FIFOMode = DMA_FIFOMode_Disable;
-	DMA_InitStructure.DMA_FIFOThreshold = DMA_FIFOThreshold_HalfFull;
-	DMA_InitStructure.DMA_MemoryBurst = DMA_MemoryBurst_Single;
-	DMA_InitStructure.DMA_PeripheralBurst = DMA_PeripheralBurst_Single;
-
-	//setting circular mode
-	DMA_InitStructure.DMA_Mode = DMA_Mode_Circular;
-	//medium priority
-	DMA_InitStructure.DMA_Priority = DMA_Priority_High;
-	//source and destination data size word=32bit
-	DMA_InitStructure.DMA_PeripheralDataSize = DMA_MemoryDataSize_Byte;
-	DMA_InitStructure.DMA_MemoryDataSize = DMA_MemoryDataSize_Byte;
-	//automatic memory destination increment enable.
-	DMA_InitStructure.DMA_MemoryInc = DMA_MemoryInc_Enable;
-	//source address increment disable
-	DMA_InitStructure.DMA_PeripheralInc = DMA_PeripheralInc_Disable;
-	//Location assigned to peripheral register will be source
-	DMA_InitStructure.DMA_DIR = DMA_DIR_MemoryToPeripheral ;
-	//chunk of data to be transfered
-	DMA_InitStructure.DMA_BufferSize = ARRAYSIZE;
-	//source and destination start addresses
-	DMA_InitStructure.DMA_PeripheralBaseAddr = (uint32_t)&USART2->DR;
-	DMA_InitStructure.DMA_Memory0BaseAddr = (uint32_t)&ADC_values[0]; //should still be the same from DMA2
-	//send values to DMA registers
-	DMA_Init(DMA1_Stream6, &DMA_InitStructure);
-	// Enable DMA2 Channel Transfer Complete interrupt
-	DMA_ITConfig(DMA1_Stream6, DMA_IT_TC, ENABLE);
-	DMA_Cmd(DMA1_Stream6, ENABLE);
-}
-
-void NVICInit(void){
-	NVIC_InitTypeDef NVIC_InitStructure;
-	//Enable DMA2 channel IRQ Channel */
-	NVIC_InitStructure.NVIC_IRQChannel = DMA2_Stream0_IRQn;
-	NVIC_InitStructure.NVIC_IRQChannelPreemptionPriority = 0;
-	NVIC_InitStructure.NVIC_IRQChannelSubPriority = 0;
-	NVIC_InitStructure.NVIC_IRQChannelCmd = ENABLE;
-	NVIC_Init(&NVIC_InitStructure);
-
-	//Enable UART channel IRQ Channel */
-	NVIC_InitStructure.NVIC_IRQChannel = DMA2_Stream6_IRQn;// USART2_IRQn;
-	NVIC_InitStructure.NVIC_IRQChannelPreemptionPriority = 0;
-	NVIC_InitStructure.NVIC_IRQChannelSubPriority = 1;
-	NVIC_InitStructure.NVIC_IRQChannelCmd = ENABLE;
-	NVIC_Init(&NVIC_InitStructure);
-}
-
-void UARTInit(void){
-	RCC_APB2PeriphClockCmd(RCC_APB1Periph_USART2, ENABLE);
-	/**
-	 * Set Baudrate to value you pass to function
-	 * Disable Hardware Flow control
-	 * Set Mode To TX and RX, so USART will work in full-duplex mode
-	 * Disable parity bit
-	 * Set 1 stop bit
-	 * Set Data bits to 8
-	 *
-	 * Initialize USART2
-	 * Activate USART2
-	 */
-	USART_InitTypeDef USART_InitStruct;
-
-	USART_InitStruct.USART_BaudRate = BaudRate;
-	USART_InitStruct.USART_HardwareFlowControl = USART_HardwareFlowControl_None;
-	USART_InitStruct.USART_Mode = USART_Mode_Tx | USART_Mode_Rx;
-	USART_InitStruct.USART_Parity = USART_Parity_No;
-	USART_InitStruct.USART_StopBits = USART_StopBits_1;
-	USART_InitStruct.USART_WordLength = USART_WordLength_8b;
-	USART_Init(USART2, &USART_InitStruct);
-
-	USART_ClockInitTypeDef USART_InitStructure;
-	USART_InitStructure.USART_Clock = USART_Clock_Disable;
-	USART_InitStructure.USART_CPOL = USART_CPOL_Low;
-	USART_InitStructure.USART_CPHA = USART_CPHA_2Edge;
-	USART_InitStructure.USART_LastBit = USART_LastBit_Disable;
-	USART_Init(USART2, &USART_InitStructure);
-
-	USART_DMACmd(USART2, (USART_DMAReq_Rx | USART_DMAReq_Tx), ENABLE);
-	USART_Cmd(USART2, ENABLE);
-}
-void DMA2_Stream0_IRQHandler(void)
-{
-	if(DMA_GetITStatus(DMA2_Stream0, DMA_IT_HTIF0))
+	while (1)
 	{
-		dma2_status = 2;
+
 	}
-	if(DMA_GetITStatus(DMA2_Stream0, DMA_IT_TCIF0))
-	{
-		dma2_status = 1;
-	}
-	DMA_ClearITPendingBit(DMA2_Stream0, DMA_IT_TCIF0 | DMA_IT_HTIF0);
-	DMA_ClearFlag(DMA2_Stream0, DMA_IT_TCIF0 | DMA_IT_HTIF0);
 }
-///*wait for DMA transfer to be done*/
-//while(dma2_status == 0) {}
-//
-//for(int i = 0, i < ARRAYSIZE, i++)
-//{
-//	//destination[i]=source[i];
-//}
+
+
